Update models namespace again
