+++ README.md
@@ -56,10 +56,10 @@ $order = $client->orders()->get(123456);
 For example update the order status and the customer email
 
 ```php
-$patch = (new \JacobDeKeizer\Ccv\Models\Internal\Resource\Orders\Patch())
+$patch = (new \JacobDeKeizer\Ccv\Models\Orders\Orders\Patch())
     ->setStatus(6)
     ->setCustomer(
-        (new \JacobDeKeizer\Ccv\Models\Internal\Entity\Personalinfo\Input)
+        (new \JacobDeKeizer\Ccv\Models\Orders\Personalinfo\Input)
             ->setEmail('example@example.com')
     );
 
@@ -69,7 +69,7 @@ $client->orders()->update(123456, $patch);
 ### Create order
 
 ```php
-$order = (new \JacobDeKeizer\Ccv\Models\Internal\Resource\Orders\Post())
+$order = (new \JacobDeKeizer\Ccv\Models\Orders\Orders\Post())
     ->setInvoicenumber(123456);
     //->set..
