+++ src/Endpoints/OrderrowsEndpoint.php
@@ -11,7 +11,7 @@ class OrderrowsEndpoint extends BaseEndpoint
     /**
      * @throws CcvShopException
      */
-    public function all(int $orderId, ?All $payload = null): Models\Resource\Collection\Orderrows
+    public function all(int $orderId, ?All $payload = null): Models\Collection\Orderrows
     {
         if ($payload === null) {
             $payload = new All();
@@ -22,6 +22,6 @@ class OrderrowsEndpoint extends BaseEndpoint
             sprintf('orders/%s/orderrows%s', $orderId, $payload->toBuilder()->toQueryString())
         );
 
-        return Models\Resource\Collection\Orderrows::fromArray($result);
+        return Models\Collection\Orderrows::fromArray($result);
     }
 }
