+++ src/Models/Orders/Personalinfo/Deliveryaddress.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace JacobDeKeizer\Ccv\Models\Orders\Entity\Personalinfo;
+namespace JacobDeKeizer\Ccv\Models\Orders\Personalinfo;
 
 use JacobDeKeizer\Ccv\Contracts\Model;
 use JacobDeKeizer\Ccv\Traits\FromArray;
 use JacobDeKeizer\Ccv\Traits\ToArray;
 
-class Billingaddress implements Model
+class Deliveryaddress implements Model
 {
     use FromArray, ToArray;
 
@@ -21,17 +21,7 @@ class Billingaddress implements Model
      private $initials;
 
     /**
-     * @var string|null First name
-     */
-     private $firstName;
-
-    /**
-     * @var string|null Last name
-     */
-     private $lastName;
-
-    /**
-     * @var string|null Deprecated. This field will no longer be used. Please use 'first_name' and 'last_name' instead.
+     * @var string|null Full name
      */
      private $fullName;
 
@@ -96,17 +86,12 @@ class Billingaddress implements Model
      private $telephone;
 
     /**
-     * @var string|null Fax number
-     */
-     private $fax;
-
-    /**
      * @var string|null Mobile phone number
      */
      private $mobile;
 
     /**
-     * @var string|null Comment left for the merchant
+     * @var string|null Comment left for the recepient
      */
      private $comment;
 
@@ -135,23 +120,7 @@ class Billingaddress implements Model
     }
 
     /**
-     * @return string|null First name
-     */
-    public function getFirstName(): ?string
-    {
-        return $this->firstName;
-    }
-
-    /**
-     * @return string|null Last name
-     */
-    public function getLastName(): ?string
-    {
-        return $this->lastName;
-    }
-
-    /**
-     * @return string|null Deprecated. This field will no longer be used. Please use 'first_name' and 'last_name' instead.
+     * @return string|null Full name
      */
     public function getFullName(): ?string
     {
@@ -255,14 +224,6 @@ class Billingaddress implements Model
     }
 
     /**
-     * @return string|null Fax number
-     */
-    public function getFax(): ?string
-    {
-        return $this->fax;
-    }
-
-    /**
      * @return string|null Mobile phone number
      */
     public function getMobile(): ?string
@@ -271,7 +232,7 @@ class Billingaddress implements Model
     }
 
     /**
-     * @return string|null Comment left for the merchant
+     * @return string|null Comment left for the recepient
      */
     public function getComment(): ?string
     {
@@ -301,29 +262,7 @@ class Billingaddress implements Model
     }
 
     /**
-     * @param string|null First name
-     * @return self
-     */
-    public function setFirstName(?string $firstName): self
-    {
-        $this->firstName = $firstName;
-        $this->propertyFilled('firstName');
-        return $this;
-    }
-
-    /**
-     * @param string|null Last name
-     * @return self
-     */
-    public function setLastName(?string $lastName): self
-    {
-        $this->lastName = $lastName;
-        $this->propertyFilled('lastName');
-        return $this;
-    }
-
-    /**
-     * @param string|null Deprecated. This field will no longer be used. Please use 'first_name' and 'last_name' instead.
+     * @param string|null Full name
      * @return self
      */
     public function setFullName(?string $fullName): self
@@ -466,17 +405,6 @@ class Billingaddress implements Model
     }
 
     /**
-     * @param string|null Fax number
-     * @return self
-     */
-    public function setFax(?string $fax): self
-    {
-        $this->fax = $fax;
-        $this->propertyFilled('fax');
-        return $this;
-    }
-
-    /**
      * @param string|null Mobile phone number
      * @return self
      */
@@ -488,7 +416,7 @@ class Billingaddress implements Model
     }
 
     /**
-     * @param string|null Comment left for the merchant
+     * @param string|null Comment left for the recepient
      * @return self
      */
     public function setComment(?string $comment): self
