+++ src/Models/Orders/Resource/Ordernotifications.php
@@ -1,17 +1,17 @@
 <?php
 
-namespace JacobDeKeizer\Ccv\Models\Orderrows\Webshop\Resource;
+namespace JacobDeKeizer\Ccv\Models\Orders\Resource;
 
 use JacobDeKeizer\Ccv\Contracts\Model;
 use JacobDeKeizer\Ccv\Traits\FromArray;
 use JacobDeKeizer\Ccv\Traits\ToArray;
 
-class Parent implements Model
+class Ordernotifications implements Model
 {
     use FromArray, ToArray;
 
     /**
-     * @var string|null Link to order
+     * @var string Link to the collection.
      */
      private $href;
 
@@ -24,18 +24,18 @@ class Parent implements Model
     }
 
     /**
-     * @return string|null Link to order
+     * @return string Link to the collection.
      */
-    public function getHref(): ?string
+    public function getHref(): string
     {
         return $this->href;
     }
 
     /**
-     * @param string|null Link to order
+     * @param string Link to the collection.
      * @return self
      */
-    public function setHref(?string $href): self
+    public function setHref(string $href): self
     {
         $this->href = $href;
         $this->propertyFilled('href');
